chore: Split logic into several files
